# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: T-Collab API
  version: 1.0.0
paths:
...
  /projects/{project_id}/invites:
    post:
      summary: Пригласить пользователя в проект
      tags:
        - Invites
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: user_email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '201':
          description: Приглашение отправлено
        '400':
          description: Пользователь уже в проекте, или приглашение уже было
            отправлено, или неверные данные
        '401':
          description: Не авторизован
        '403':
          description: Недостаточно прав для приглашения пользователя
        '404':
          description: Проект или пользователь не найден
    get:
      summary: Получить все приглашения в этом проекте
      tags:
        - Invites
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Приглашения получены
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invite'
                required: [data]
        '400':
          description: Неверные данные
        '401':
          description: Не авторизован
        '403':
          description: Недостаточно прав для просмотра приглашений
        '404':
          description: Проект не найден
  /invites:
    get:
      summary: Получить список приглашений пользователя (себя)
      tags:
        - Invites
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Приглашения получены
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invite'
                required: [data]
        '401':
          description: Не авторизован
  /invites/{invite_id}:
    delete:
      summary: Принять или отклонить приглашение
      tags:
        - Invites
      security:
        - BearerAuth: []
      parameters:
        - name: invite_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: accept
          in: query
          required: true
          description: Прниять или отклонить запрос
          schema:
            type: boolean
      responses:
        '200':
          description: Successful
        '400':
          description: Неверные данные
        '401':
          description: Не авторизован
        '403':
          description: Недостаточно прав для изменения приглашения
        '404':
          description: Приглашение не найдено
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    googleOAuth:
      type: oauth2
      description: Авторизация через Google OAuth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            email: Доступ к email пользователя
            profile: Доступ к имени пользователя
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
...
    Invite:
      type: object
      properties:
        invite_id:
          type: string
          format: uuid
          description: Идентификатор приглашения
        project_id:
          type: string
          format: uuid
          description: Идентификатор проекта
        email:
          type: string
          format: email
          description: Email приглашённого пользователя
        invited_at:
          type: string
          format: date-time
          description: Дата и время отправки приглашения
        sender_nickname:
          type: string
        project_name:
          type: string
        receiver_nickname:
          type: string
        receiver_avatar:
          type: string
      required:
        - invite_id
        - project_id
        - email
        - invited_at
        - project_name
        - sender_nickname
        - receiver_nickname
      additionalProperties: false